# gateway/src/main/resources/static/openapi.yaml
openapi: 3.0.1
info:
  title: Larpex
  description: Larpex API
  version: 0.0.1
servers:
  - url: /
paths:
...
  /interactions:
    post:
      tags:
        - interaction-controller
      operationId: postInteract
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InteractCommand"
        required: true
      responses:
        "400":
          description: Interaction not performed due to malformed qr code
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/InteractResponse"
        "200":
          description: Interaction performed
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/InteractResponse"
components:
  schemas:
...
    InteractCommand:
      type: object
      properties:
        userId:
          type: string
          format: uuid
        qrCode:
          type: string
    InteractResponse:
      type: object
      properties:
        interactionType:
          type: string
          enum:
            - OPEN_DOOR
            - OPEN_CHEST
...